+++ packages/signers/ledger-signer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Use right signed extensions for transaction building
+
 ## 0.2.13 to 0.2.14 - 2026-01-05
 
 ### Fixed